fix course service and user service

+++ user-service/src/main/java/il/ac/afeka/fdp/user/infra/UserServiceImpl.java
@@ -55,6 +55,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUserById(String id) {
+        if (!this.repository.existsById(id)) {
+            throw new UserNotFound();
+        }
         this.repository.deleteById(id);
     }
